Changed every weight from int to double

+++ src/structures/Algorithms.java
@@ -108,8 +108,8 @@ public class Algorithms {
 	 * @param v initial vertex
 	 * @return Array with the shortest paths form the initial vertex
 	 */
-	public static <V> int[] dijkstra(int[][] g, int v){
-		int[] dis = new int[g.length];
+	public static <V> double[] dijkstra(double[][] g, int v){
+		double[] dis = new double[g.length];
 		boolean[] vis = new boolean[g.length];
 		
 		
@@ -139,8 +139,8 @@ public class Algorithms {
 	 * @param vis array of vertices visited
 	 * @return Index of the minimum value
 	 */
-	private static int minIndex(int[] dis, boolean[] vis) { 
-        int min = Integer.MAX_VALUE;
+	private static int minIndex(double[] dis, boolean[] vis) { 
+		double min = Integer.MAX_VALUE;
         int indx = -1; 
   
         for (int i=0; i<dis.length; i++) 
@@ -158,10 +158,10 @@ public class Algorithms {
 	 * @param w Weighted matrix
 	 * @return Matrix with the shortest path form the vertices to each other
 	 */
-	public static <V> int[][] floydWarshall(int[][] w){
+	public static <V> double[][] floydWarshall(double[][] w){
 		int n = w.length;
-		int[][] D = w;
-		int v = 0;
+		double[][] D = w;
+		double v = 0;
 		
 		for(int k=0; k<n; k++) {
 			for(int i=0; i<n; i++) {
@@ -190,11 +190,11 @@ public class Algorithms {
 	 * @param w Weighted matrix
 	 * @return Matrix showing the connections of the MST
 	 */
-	public static int[][] Kruskal(int[][] w){
+	public static double[][] Kruskal(double[][] w){
 		
 		DisjointSet<Integer> set = new DisjointSet<>();
 		
-		int[][] MST = new int[w.length][w.length];
+		double[][] MST = new double[w.length][w.length];
 
 		for(int i = 0; i < w.length; i++) {
 			set.makeSet(i);
@@ -204,8 +204,8 @@ public class Algorithms {
 		class obj {
 			int A;
 			int B;
-			int P;			
-			obj(int a, int b, int p){
+			double P;			
+			obj(int a, int b, double p){
 				A = a;
 				B = b;
 				P = p;
@@ -216,7 +216,7 @@ public class Algorithms {
 			int getB() {
 				return B;
 			}
-			int getP() {
+			double getP() {
 				return P;
 			}
 		}
@@ -225,7 +225,7 @@ public class Algorithms {
 		ArrayList<obj> aristas = new ArrayList<>();		
 		for(int i = 0; i < w.length;  i++) {
 			for(int j = 0; j < w.length; j++) {
-				int peso = w[i][j];
+				double peso = w[i][j];
 				if(peso > 0 && peso < Integer.MAX_VALUE) {
 					obj o = new obj(i, j, peso);
 					aristas.add(o);
@@ -263,9 +263,9 @@ public class Algorithms {
 	 * @param matrix Weighted matrix
 	 * @return array with the weight of the MST
 	 */
-	public static int[] prim(int[][] matrix){
-		int[] mst = new int[matrix.length];
-		int[] weight = new int[matrix.length];
+	public static double[] prim(double[][] matrix){
+		double[] mst = new double[matrix.length];
+		double[] weight = new double[matrix.length];
 		boolean[] inMst = new boolean[matrix.length];
 		
 		for (int i = 0; i < matrix.length; i++) {
